backend/requirements.txt
flask==3.0.0
flask-cors==6.0.2
python-dotenv==1.0.0
openai==1.99.1
Pillow==12.0.0
torch>=2.1.0
torchvision>=0.16.0
pycocotools==2.0.7

# SAM3 specific dependencies
# Add sam3 package or install from source
git+https://github.com/facebookresearch/sam3.git
opencv-python # 自动找配合 sam3 的版本
psutil
